Revert experimental 1-to-1 PE-MIG connection code
For a number of PEs smaller than or equal to the
number of available MIGs, this code applied a
1-to-1 connection between PE(s) and MIG(s),
removing the "out" AXI interconnect.

It was done in a quick & dirty style, only meant
to be used for benchmarking. Still, it might be
useful to someone at some point, thus I'm
not just dropping the commits that introduced
this code.