PEs are now always wrapped before post-pe-create plugin
The full_axi_slave_wrapper post-pe-create plugin was
wrapping a PE to insert a AXI lite to AXI full converter.
This creates a problem for other post-pe-create plugins.
Especially since the order in which post-pe-create
plugins are executed is random.

This commit changes that a PE is now always wrapped. All
post-pe-create plugins now have the same starting
conditions. They also always have to return the PE so that
following plugins have access to them.