The wrapper around PEs is now created with connected clk and rst
The goal is to make post-pe-create plugins independent of execution
order. With this commit the wrapper around the PE is created
with clk and rst ports to the outside. This means that
the full_axi_slave_wrapper does not have to execute at all.
Now it is only required when the PE has an AXI full port.