Add `libtapasco_tests` to CMake build config
In order to detect when the compilation of `libtapasco_tests` fails, I
want to add it to the CI system.

First, I thought about adding it directly to the `tapasco-build-libs`
script but since it's just a thin wrapper around KBuild and CMake I
decided to try to add it to CMake instead. Compiling Rust with CMake
seems not to be possible in a straightforward manner, so I then added a
custom target which calls Cargo to build it because Cargo handles
everything just nicely.
The `CMakeLists.txt` of `libtapasco` was not applicable here because in
order for `libtapasco_tests` to show up under the executable directory
it needs to be added as an executable, not as a library as in
`libtapasco`.
This makes another workaround necessary to overwrite the linker with a
`/usr/bin/true` no-op because CMake really cannot be persuaded to just
not link an executable but Cargo has already run the linker.

Note that Cross compilation is not supported this way.