Switch to write lock and add second try
Hold the mmap_write_lock() instead of mmap_read_lock() for host-to-device migrations to prevent that the CPU and device page fault handler run at the same time. During concurrent page accesses causing a Ping-Pong effect migrations to device failed and most accelerators do not handle the return SLVERRs correctly.

If the migration of pages fail for ODPMs in host-to-device direction, try it a second time (this may happen if a race for a zero-page between device and host occurs).