Optimize device memory free
Free contiguous blocks of device memory at the end of the migration routines.
This way we prevent one-by-one freeing by the "page_free" system call.
However to prevent "double freeing" we need to mark currently allocated pages by letting the 'zone_device_data' field to the device.
The pointer is nulled when freeing the pages during migration.
'page_free' checks for this pointer and only initiates a free if the 'zone_device_data' filed is nun-NULL.