Configure MAC during compose
Add an additional option 'mac_addr' during tapasco compose if network migrations are enabled.
The given MAC address is set in the NetworkPageDMA (which is now configurable) AND 100G Ethernet core as source MAC for pause packets.
The driver reads the respective MAC address from the NetworkPageDMA core during initialization and saves it for later use.