Add PCIe E2E transfers
Add endpoint-to-endpoint transfers over PCIe as additional option for direct device-to-device page migrations.
We use a second BAR to expose the device memory to the PCIe bus for that. The PageDMA core of the source device can then write the pages directly into memory on the destination device.

TODO: HW support must be scripted to use this.