Optimize device page free
For dev-to-dev and dev-to-ram migrations, do not create a work struct for every contiguous memory range to be freed,
but create one work struct with the source pages array. This speeds up the migration a lot if the source pages are
scattered in memory, since we do not have to create 100s/1000s of work structs which are then handled one by one by
worker threads.