Add first mockup of `tapasco-debug` in Rust
This is the first commit for a work-in-progress effort on Issue #264.

Since `tapasco-debug` was removed in the v2020.10 release because the
ncurses-based UI was "quite flaky" you had to use two different Tapasco
workspaces to have the debugger still around while using the recent
Tapasco version.
This reimplementation of `tapasco-debug` is written in Rust and
uses the `libtui-rs` and `crossterm` crates for the terminal user
interface along with the new Tapasco runtime.

For the moment it provides a list of PEs which can be selected to see
the first 16 argument registers along with some generic bitstream and
platform information.
Writing to argument registers, starting PEs, dumping Local Memory is not
possible yet.

To integrate it into the build process of this repository it uses the
same mechanism to build using Cargo from CMake as PR #285.