Implement Monitor, Debug and Unsafe Mode
As discussed in the respective PR #290
(https://github.com/esa-tu-darmstadt/tapasco/pull/290)
there are now three differenct access modes: Monitor, Debug, Unsafe.

This commit introduces command-line subcommands to choose the mode in
which to operate.
In `Monitor` mode nothing can be modified only another runtime
can be monitored.
In `Debug` Mode the argument registers can also be set interactively.
In `Unsafe` Mode the only difference to `Debug` Mode is that the driver
is released, so another host application can run in parallel.