Add list of acquired devices to miscdev
Track all opened devices in a list. This is required if an application uses more than one device in parallel. In the current implementation only the last acquired device is saved in the miscdev, and when closing the device file in the user-space runtime, only acceses to this single device is released. By tracking all acquired devices in a list, all devices are released properly when closing the miscdev.