Create VFIO mapping during allocate
Move creation of VFIO mapping from ```copy_to()``` to ```allocate```.
Otherwise ```makeOutOnly``` will fail since in this case ```copy_to``` is never called. Now ```makeInOnly``` and ```makeOutOnly``` annotations of buffers are just ignored, since they do not have influence anyway as we are not moving any data thanks to the use of VFIO and SMMU mappings.
Closes #341.