Add non-blocking calls to acquire and release jobs
The function ```try_acquire_pe()``` of a device allows probing whether a PE is available without blocking the current thread.
It immediately returns an empty Option if no PE is available, or the acquired Job otherwise.

Similarly ```try_release()``` on a Job immediately returns an empty Option if the job is still running, or
the return value and copy backs if it has finished.

Non-blocking equivalent of functions called for acquiring and releasing jobs are added respectively.