Add non-blocking launch and release to C++ API
The function callback of ```job_future``` is wrapped in an object. This allows a variant with
argument setting whether the callback blocks till job completion or returns immediately with an
error code if the job is not finished yet. For backwards compatibility the always blocking variant
without boolean argument still exists.

```try_launch()``` does not blocking the calling thread and immediately returns an error code
if no PE is available. A JobFuture object is passed by reference and the respective callback
is set if a PE could be acquired.

Respective functions are added to the FFI using the underlying functions in the Rust API.