Adjust SVM extension on U50
Adjust the SVM extension to the new version of U50 support.
SVM cannot use multiple HBM ports due to the single IOMMU. So we have to introduce
an interconnect tree if we have multiple PE masters. Also the address is adjusted
to prevent artifacts.
DMA and PE access still use different HBM ports.