Locate CIPS in host subsystem
CIPS and NoC must be generated together in host subsystem, but NoC is moved
later to memory subsystem.
Do not throw an error if an interface is not specified during address map
generation. Due to the NoC some address segments are already specified by
other interfaces and respecifying leads to conflicts.