Add QDMA streaming support to runtime
Streaming transfers work in principle as other DMA transfers. However, as they
are executed during runtime of the PE, a concurrent thread is launched handling
the transfers. After the PE has finished we wait for joining of these DMA threads.
A streaming parameter does not initiate any write to a control register.