Free job object in ffi interface
Job object is freed after the job has completed, when using the C++ API.
The job object hold a reference to the scheduler. This has prevented to
destroy the Tapasco object properly (e.g. interrupts where not deregistered
and DMA buffers not deallocated). Caused errors while trying to re-open the device.